+++ src/Wambo/ComposerInstaller/Exception/InvalidArgumentException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Wambo\ComposerInstaller\Exception;
+
+class InvalidArgumentException extends \InvalidArgumentException
+{
+}