+++ src/Wambo/ComposerInstaller/Plugin.php
@@ -6,10 +6,22 @@ use Composer\Composer;
 use Composer\IO\IOInterface;
 use Composer\Plugin\PluginInterface;
 
+/**
+ * Class Plugin
+ *
+ * @package Wambo\ComposerInstaller
+ */
 class Plugin implements PluginInterface
 {
 
-    public function activate(Composer $composer, IOInterface $io){
+    /**
+     * Add The WamboInstaller to compser
+     *
+     * @param Composer    $composer
+     * @param IOInterface $io
+     */
+    public function activate(Composer $composer, IOInterface $io)
+    {
         $installer = new WamboInstaller($io, $composer);
         $composer->getInstallationManager()->addInstaller($installer);
     }