+++ src/Wambo/ComposerInstaller/WamboInstaller.php
@@ -4,6 +4,12 @@ namespace Wambo\ComposerInstaller;
 use Composer\Installer\LibraryInstaller;
 use Composer\Package\PackageInterface;
 use Composer\Repository\InstalledRepositoryInterface;
+use League\Flysystem\Adapter\Local;
+use League\Flysystem\Filesystem;
+use Wambo\Core\Module\JSONModuleStorage;
+use Wambo\Core\Module\Module;
+use Wambo\Core\Module\ModuleMapper;
+use Wambo\Core\Module\ModuleRepository;
 
 /**
  * WamboInstaller is an composer installer to add a "wambo-module" to a
@@ -15,8 +21,12 @@ class WamboInstaller extends LibraryInstaller
 {
 
     const PACKAGE_TYPE = 'wambo-module';
+    const EXTRA_BOOTSTRAP_CLASS_KEY = 'class';
+    const AUTOLOAD_TYPE = 'psr-4';
 
-    const MODULES_JSON_PATH = 'config/wambo/modules.json';
+    const MODULES_JSON_FILENAME = 'modules.json';
+
+    private $moduleRepository;
 
     /**
      * Decides if the installer supports the given type
@@ -37,36 +47,31 @@ class WamboInstaller extends LibraryInstaller
      */
     public function install(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
+        $this->bootstrap();
+
         // display info on "composer update" on command line
         echo 'add  ' . $package->getName() . ' to wambo modules';
         echo PHP_EOL;
 
-        $autoload = $package->getAutoload();
-        if(array_key_exists('psr-4', $autoload)){
-            $namespaces = array_keys($autoload['psr-4']);
-            if(!empty($namespaces)){
-                $namespace_parts = array_filter(preg_split('/\\\\/', array_pop($namespaces)));
-                $module_classname = $namespace_parts[count($namespace_parts) -1];
+        $wamboInstallerService = new WamboInstallerService($package);
 
-                if( file_exists(self::MODULES_JSON_PATH) ){
-                    $modules_json = file_get_contents(self::MODULES_JSON_PATH);
-                    $modules = json_decode($modules_json, true);
-                }else{
-                    $modules = array();
-                }
+        $className = $wamboInstallerService->getBootstrapClassName();
+        $name = $package->getName();
+        $version = $package->getVersion();
 
-
-                $module_entity = array(
-                    'name' => $module_classname,
-                    'namespace' => implode('\\', $namespace_parts)
-                );
-                array_push($modules, $module_entity);
-
-                $modules = array_unique($modules, SORT_REGULAR);
-                file_put_contents(self::MODULES_JSON_PATH, json_encode($modules, JSON_PRETTY_PRINT));
-            }
-        }
+        $module = new Module($name, $version, $className);
+        $this->moduleRepository->add($module);
 
         parent::install($repo, $package);
+    }
+
+    private function bootstrap()
+    {
+        $filesystemAdapter = new Local($this->vendorDir);
+        $filesystem = new Filesystem($filesystemAdapter);
+        $mapper = new ModuleMapper();
+        $storage = new JSONModuleStorage($filesystem, self::MODULES_JSON_FILENAME);
+
+        $this->moduleRepository = new ModuleRepository($storage, $mapper);
     }
 }